+++ R/ext2spy.R
@@ -3,30 +3,36 @@
 #' @description 
 #' Convert a spatial extent to polygons.
 #' 
-#' @param x An `Extent` object, or any object from which an `Extent` can be 
-#'   extracted, e.g. `Raster*`.
-#' @param crs Coordinate reference system passed to [sp::proj4string()].
+#' @param x A `SpatExtent` object, or any object from which such an object can 
+#'   be extracted, e.g. `SpatRaster`.
+#' @param crs Coordinate reference system set via [terra::crs()].
 #' @param as_sf `logical`. If `TRUE` (default), the returned object is of class 
 #'   `sf` rather than `Spatial*`.
 #' 
-#' @return Depending on 'as_sf', either a `sf` or `SpatialPolygons` object.
+#' @return Depending on 'as_sf', either a `c(sf, data.frame)` or `SpatVector` 
+#'   object.
 #' 
 #' @author Florian Detsch
 #' 
-#' @seealso [raster::extent()].
+#' @seealso [terra::ext()].
 #' 
 #' @examples 
-#' ext = extent(c(25, 70, -5, 30))
+#' ext = terra::ext(c(25, 70, -5, 30))
 #' ext2spy(ext) # 'sf' (default)
 #' ext2spy(ext, as_sf = FALSE) # 'Spatial*'
-#'
+#' 
+#' @importFrom sf st_as_sf
+#' @importFrom terra as.polygons crs ext
+#' 
 #' @export
-ext2spy = function(x, crs = "+init=epsg:4326", as_sf = TRUE) {
-  if (!inherits(x, "Extent"))
-    x = raster::extent(x)
-    
-  spy = as(x, "SpatialPolygons")
-  sp::proj4string(spy) = crs
+ext2spy = function(x, crs = "EPSG:4326", as_sf = TRUE) {
+  
+  if (!inherits(x, "SpatExtent")) {
+    x = terra::ext(x)
+  }
+  
+  spy = terra::as.polygons(x)
+  terra::crs(spy) = crs
   
   if (as_sf) {
     spy = sf::st_as_sf(spy)