+++ R/latticeCombineGrid.R
@@ -5,7 +5,7 @@
 #' Note that the global plot settings (e.g. 'xlim', 'ylim', ...) are taken from 
 #' the first object though the user can specify whether 'scales' should be 
 #' identical or not. This is particularly useful when looping over large amounts
-#' of data using [lapply()] (see Examples).
+#' of data using [lapply()] or the like (see Examples).
 #' 
 #' @param trellis.list A `list` containing **lattice** plot objects.
 #' @param between Space between panels.
@@ -22,35 +22,44 @@
 #' [latticeExtra::c.trellis()].
 #' 
 #' @examples
-#' #load data
-#' #Use a probability map assuming high potential for city expansion is just 
-#' #resulting from proximity to current urban area:
-#' pred <- raster(system.file("extdata/probability.rst", package = "Orcs"))
+#' library(lattice)
 #' 
-#' #observed city growth between 1990 and 2006
-#' obs <- raster(system.file("extdata/citygrowth.tif", package = "Orcs"))
+#' p1 = xyplot(1:10 ~ 1:10)
+#' p2 = xyplot(10:1 ~ 1:10)
 #' 
-#' #masking current urban area since these pixels have no potential for change
-#' mask <- raster(system.file("extdata/citymask.tif", package = "Orcs"))
+#' (
+#'   p = latticeCombineGrid(
+#'     list(p1, p2)
+#'   )
+#' )
 #' 
-#' #create data list
-#' dat <- list(pred, obs, mask)
-#' 
-#' #create list of lattice plots
-#' plist <- lapply(seq(dat), function(i) {
-#'   spplot(dat[[i]], scales = list(draw = TRUE))
-#' })
-#' 
-#' \donttest{
-#' #draw individually
-#' plist[[1]]
-#' plist[[2]]
-#' plist[[3]]
+#' if (requireNamespace("raster", quietly = TRUE)) {
+#'   # load data
+#'   # Use a probability map assuming high potential for city expansion is just 
+#'   # resulting from proximity to current urban area:
+#'   prd = raster::raster(system.file("extdata/probability.rst", package = "Orcs"))
+#'   
+#'   # observed city growth between 1990 and 2006
+#'   obs = raster::raster(system.file("extdata/citygrowth.tif", package = "Orcs"))
+#'   
+#'   # masking current urban area since these pixels have no potential for change
+#'   msk = raster::raster(system.file("extdata/citymask.tif", package = "Orcs"))
+#'   
+#'   # create data list
+#'   dat <- list(prd, obs, msk)
+#'   
+#'   # create list of lattice plots
+#'   plist <- lapply(dat, raster::spplot, scales = list(draw = TRUE))
+#'   
+#'   # # draw individually
+#'   # plist[[1]]
+#'   # plist[[2]]
+#'   # plist[[3]]
+#'     
+#'   # combine to grid, using c(1, 3) layout
+#'   p = latticeCombineGrid(plist, layout = c(1, 3))
+#'   print(p)
 #' }
-#' 
-#' #combine to grid, using c(1, 3) layout
-#' p <- latticeCombineGrid(plist, layout = c(1, 3))
-#' print(p)
 #' 
 #' @export
 latticeCombineGrid <- function(trellis.list,