+++ inst/tinytest/test-ifMissing.R
@@ -1,25 +1,38 @@
 ## raster files
-logo = system.file("external/rlogo.grd", package = "raster")
+logo = system.file("ex/logo.tif", package = "terra")
 
 ofl = file.path(tmp <- tempdir(), "rlogo.tif")
-if (file.exists(ofl)) jnk = file.remove(ofl)
+
+avl = file.exists(ofl)
+if (avl) {
+  jnk = file.rename(ofl, paste0(ofl, ".bu"))
+}
 
 s = ifMissing(logo)
-expect_inherits(s, "RasterBrick")
+expect_inherits(s, "SpatRaster")
 
 s2 = ifMissing(ofl, arg1 = "filename", x = s, datatype = "INT1U")  
-expect_inherits(s2, "RasterBrick")
+expect_inherits(s2, "SpatRaster")
 
 s3 = ifMissing(ofl)
-expect_inherits(s3, "RasterBrick")
+expect_inherits(s3, "SpatRaster")
 
 jnk = file.remove(ofl)
+
+if (avl) {
+  jnk = file.copy(paste0(ofl, ".bu"), ofl, overwrite = TRUE)
+  jnk = file.remove(paste0(ofl, ".bu"))
+}
 
 ## text files
 data(iris)
 
 ofl = file.path(tmp, "iris.csv")
-if (file.exists(ofl)) jnk = file.remove(ofl)
+
+avl = file.exists(ofl)
+if (avl) {
+  jnk = file.rename(ofl, paste0(ofl, ".bu"))
+}
 
 jnk = ifMissing(ofl, fun1 = write.csv, x = iris, file = ofl, row.names = FALSE)
 expect_null(jnk)
@@ -29,11 +42,15 @@ dat$Species = factor(dat$Species)
 expect_equal(dat, iris)
 
 jnk = file.remove(ofl)
-expect_true(jnk)
+
+if (avl) {
+  jnk = file.copy(paste0(ofl, ".bu"), ofl, overwrite = TRUE)
+  jnk = file.remove(paste0(ofl, ".bu"))
+}
 
 fun = function(x, file = "", ...) {
-  write.csv(x, file, ...)
-  read.csv(file)
+  utils::write.csv(x, file, ...)
+  utils::read.csv(file)
 }
 dat = ifMissing(ofl, fun1 = fun, arg1 = "file", x = iris, quote = FALSE, row.names = FALSE)
 dat$Species = factor(dat$Species)