+++ NEWS.md
@@ -1,3 +1,21 @@
+# Orcs 1.2.2.9002 (2022-12-22)
+
+#### ✨ features and improvements
+
+#### 🐛 bug fixes
+
+#### 💬 documentation etc
+
+  * `KiLi` built-in data is no longer available as `SpatRaster` objects do not 
+    work with `save()` (see <https://github.com/rspatial/terra/issues/549>); use
+    `terra::rast(system.file("extdata/KiLi.tif", package = "Orcs"))` instead.
+
+#### 🍬 miscellaneous
+
+  * Eliminates **rgdal** dependency (#9)
+  * Replaces **raster** related code and objects with **terra** analogs
+
+
 # Orcs 1.2.2 (2022-09-26)
 
 #### 🍬 miscellaneous